+++ queries/dmpworks/python/dmpworks/model/dmp_model.py
@@ -8,11 +8,14 @@ import pendulum
 from pydantic import BaseModel, field_serializer, field_validator
 
 from dmpworks.funders.award_id import AwardID
+from dmpworks.model.common import Author, Funder, Institution, to_camel
 
 
 class DMPModel(BaseModel):
     model_config = {
         "arbitrary_types_allowed": True,
+        "alias_generator": to_camel,
+        "populate_by_name": True,
     }
 
     doi: str
@@ -50,11 +53,6 @@ class DMPModel(BaseModel):
         return v.to_date_string()
 
 
-class Funder(BaseModel):
-    name: Optional[str]
-    id: Optional[str]
-
-
 class FundingItem(BaseModel):
     funder: Optional[Funder]
     funding_opportunity_id: Optional[str]
@@ -62,23 +60,12 @@ class FundingItem(BaseModel):
     award_id: Optional[str]
 
 
-class Institution(BaseModel):
-    name: Optional[str]
-    ror: Optional[str]
-
-
-class Author(BaseModel):
-    orcid: Optional[str]
-    first_initial: Optional[str]
-    given_name: Optional[str]
-    middle_initials: Optional[str]
-    middle_names: Optional[str]
-    surname: Optional[str]
-    full: Optional[str]
-
-
 class ExternalData(BaseModel):
-    model_config = {"arbitrary_types_allowed": True}
+    model_config = {
+        "arbitrary_types_allowed": True,
+        "alias_generator": to_camel,
+        "populate_by_name": True,
+    }
 
     updated: pendulum.DateTime
     awards: list[Award]
@@ -98,7 +85,11 @@ class ExternalData(BaseModel):
 
 
 class Award(BaseModel):
-    model_config = {"arbitrary_types_allowed": True}
+    model_config = {
+        "arbitrary_types_allowed": True,
+        "alias_generator": to_camel,
+        "populate_by_name": True,
+    }
 
     funder: Optional[Funder]
     award_id: Optional[AwardID]