+++ queries/dmpworks/python/dmpworks/opensearch/dmp_works.py
@@ -1,16 +1,16 @@
-import dataclasses
-import json
 import logging
+import math
 import pathlib
 from typing import Callable, Optional
 
 import jsonlines
+import pendulum
 from opensearchpy import OpenSearch
 from tqdm import tqdm
 
-from dmpworks.model.dmp_model import DMPModel
+from dmpworks.model.dmp_model import Award, DMPModel
+from dmpworks.model.related_work_model import ContentMatch, DoiMatch, DoiMatchSource, ItemMatch, RelatedWork
 from dmpworks.model.work_model import WorkModel
-from dmpworks.opensearch.explanations import explain_match, Group
 from dmpworks.opensearch.utils import make_opensearch_client, OpenSearchClientConfig, yield_dmps
 from dmpworks.utils import timed
 
@@ -38,9 +38,15 @@ def dmp_works_search(
     if parallel_search and include_named_queries_score:
         log.warning("Unable to use include_named_queries_score with msearch, query scores will not be returned.")
 
-    def write_works(works: list[DMPWorksSearchResult], count: int):
+    def write_works(works: list[RelatedWork], count: int):
         for work in works:
-            writer.write(work.to_dict())
+            writer.write(
+                # Convert fields to CamelCase for database
+                work.model_dump(
+                    by_alias=True,
+                    mode="json",
+                )
+            )
         pbar.update(count)
 
     with tqdm(total=0, desc="Find DMP work matches with OpenSearch", unit="doc") as pbar:
@@ -94,24 +100,6 @@ def dmp_works_search(
                     write_works(works, len(batch))
 
 
-@dataclasses.dataclass
-class DMPWorksSearchResult:
-    dmp_doi: str
-    work_doi: str
-    score: float
-    work: WorkModel
-    explanations: dict[str, Group]
-
-    def to_dict(self) -> dict:
-        return {
-            "dmp_doi": self.dmp_doi,
-            "work_doi": self.work_doi,
-            "score": self.score,
-            "work": self.work.model_dump(),
-            "explanations": {name: group.to_dict() for name, group in self.explanations.items()},
-        }
-
-
 def msearch_dmp_works(
     client: OpenSearch,
     index_name: str,
@@ -120,7 +108,7 @@ def msearch_dmp_works(
     project_end_buffer_years: int = 3,
     max_concurrent_searches: int = 125,
     max_concurrent_shard_requests: int = 12,
-) -> list[DMPWorksSearchResult]:
+) -> list[RelatedWork]:
     # Execute searches
     body = []
     for dmp in dmps:
@@ -151,7 +139,7 @@ def search_dmp_works(
     max_results: int = 100,
     project_end_buffer_years: int = 3,
     include_named_queries_score: bool = False,
-) -> list[DMPWorksSearchResult]:
+) -> list[RelatedWork]:
     body = build_query(dmp, max_results, project_end_buffer_years)
     response = client.search(
         body=body,
@@ -162,34 +150,95 @@ def search_dmp_works(
     return collate_results(dmp, hits)
 
 
-def collate_results(dmp: DMPModel, hits: list[dict]) -> list[DMPWorksSearchResult]:
-    results: list[DMPWorksSearchResult] = []
+def parse_matched_queries(matched_queries: list[str] | dict[str, float]) -> dict[str, float]:
+    if isinstance(matched_queries, list):
+        return {key: math.nan for key in matched_queries}
+    return matched_queries
+
+
+def collate_results(dmp: DMPModel, hits: list[dict]) -> list[RelatedWork]:
+    results: list[RelatedWork] = []
     for hit in hits:
-        work_doi: str = hit.get("_id")
+        work_doi = hit.get("_id")
         score: float = hit.get("_score", 0.0)
-        work = WorkModel.model_validate(hit.get("_source", {}))
-        matched_queries = hit.get("matched_queries")
-        highlights = hit.get("highlight")
-        explanation = explain_match(dmp, work, matched_queries, highlights)
-        results.append(DMPWorksSearchResult(dmp.doi, work_doi, score, work, explanation))
+        work = WorkModel.model_validate(hit.get("_source", {}), by_name=True, by_alias=False)
+        matched_queries = parse_matched_queries(hit.get("matched_queries", []))
+        highlights = hit.get("highlight", {})
+
+        # Construct DOI match
+        doi_found = "funded_dois" in matched_queries
+        sources = []
+        if doi_found:
+            for award in dmp.external_data.awards:
+                if work_doi in award.funded_dois:
+                    parent = award.award_id
+                    awards = [parent] + award.award_id.related_awards
+                    for child in awards:
+                        if child.award_url() is not None:
+                            parent_award_id = parent.identifier_string() if parent != child else None
+                            sources.append(
+                                DoiMatchSource(
+                                    parent_award_id=parent_award_id,
+                                    award_id=child.identifier_string(),
+                                    award_url=child.award_url(),
+                                )
+                            )
+        doi_match = DoiMatch(
+            found=doi_found,
+            score=matched_queries.get("funded_dois", 0.0),
+            sources=sources,
+        )
+
+        # Construct content match (based on title and abstract)
+        title_highlights = highlights.get("title", [])
+        abstract_highlights = highlights.get("abstract", [])
+        content_score = matched_queries.get("content", 0.0)
+        content_matched = "content" in matched_queries
+        content_match = ContentMatch(
+            score=content_score,
+            title_highlight=title_highlights[0] if title_highlights and content_matched else None,
+            abstract_highlights=abstract_highlights if content_matched else [],
+        )
+
+        # Construct matches based on inner hits
+        inner_hits = hit.get("inner_hits", {})
+        author_matches = to_item_matches(inner_hits, "authors")
+        institution_matches = to_item_matches(inner_hits, "institutions")
+        funder_matches = to_item_matches(inner_hits, "funders")
+        award_matches = to_item_matches(inner_hits, "awards")
+
+        results.append(
+            RelatedWork(
+                dmp_doi=dmp.doi,
+                work=work,
+                score=score,
+                doi_match=doi_match,
+                content_match=content_match,
+                author_matches=author_matches,
+                institution_matches=institution_matches,
+                funder_matches=funder_matches,
+                award_matches=award_matches,
+            )
+        )
     return results
 
 
-################
-# Query builder
-################
-
-
-def name_group(group: str) -> str:
-    return json.dumps({"g": group, "lvl": "group"}, separators=(",", ":"))
-
-
-def name_entity(group: str, idx: int) -> str:
-    return json.dumps({"g": group, "lvl": "entity", "i": idx}, separators=(",", ":"))
-
-
-def name_value(group: str, idx: int, key: str, value: str) -> str:
-    return json.dumps({"g": group, "lvl": "value", "i": idx, "k": key, "v": value}, separators=(",", ":"))
+def to_item_matches(inner_hits: dict, hit_name: str) -> list[ItemMatch]:
+    matches = []
+    hits = inner_hits.get(hit_name, {}).get("hits", {}).get("hits", [])
+    for hit in hits:
+        offset = hit.get("_nested", {}).get("offset")
+        score = hit.get("_score")
+        matched_queries = parse_matched_queries(hit.get("matched_queries", []))
+        fields = [field.replace(f"{hit_name}.", "") for field in matched_queries.keys()]
+        matches.append(
+            ItemMatch(
+                index=offset,
+                score=score,
+                fields=fields,
+            )
+        )
+    return matches
 
 
 def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) -> dict:
@@ -200,7 +249,7 @@ def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) 
         should.append(
             {
                 "constant_score": {
-                    "_name": name_group("funded_dois"),
+                    "_name": "funded_dois",
                     "boost": 15,
                     "filter": {
                         "ids": {"values": dmp.funded_dois},
@@ -209,47 +258,11 @@ def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) 
             }
         )
 
-    # Award IDs
-    award_queries = []
-    group = "awards"
-    for idx, award in enumerate(dmp.external_data.awards):
-        queries = []
-        for award_id in award.award_id.all_variants:
-            queries.append(
-                {
-                    "constant_score": {
-                        "_name": name_value(group, idx, "award_id", award_id),
-                        "filter": {"term": {"award_ids": award_id}},
-                        "boost": 10,
-                    }
-                }
-            )
-        award_queries.append(
-            {
-                "dis_max": {
-                    "_name": name_entity(group, idx),
-                    "tie_breaker": 0,
-                    "queries": queries,
-                }
-            }
-        )
-    if len(award_queries):
-        should.append(
-            {
-                "bool": {
-                    "_name": name_group(group),
-                    "minimum_should_match": 1,
-                    "should": award_queries,
-                },
-            }
-        )
-
     # Authors
-    # Combines author_orcids and author surnames into a single feature
     authors = build_entity_query(
         "authors",
-        "author_orcids",
-        "author_names",
+        "authors.orcid",
+        "authors.full",
         dmp.authors,
         lambda author: author.orcid,
         lambda author: author.surname,
@@ -258,11 +271,10 @@ def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) 
         should.append(authors)
 
     # Institutions
-    # Combines both affiliation_rors and affiliation_names into a single feature
     institutions = build_entity_query(
         "institutions",
-        "affiliation_rors",
-        "affiliation_names",
+        "institutions.ror",
+        "institutions.name",
         dmp.institutions,
         lambda inst: inst.ror,
         lambda inst: inst.name,
@@ -271,17 +283,24 @@ def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) 
         should.append(institutions)
 
     # Funders
-    # Combines both funder_ids and funder_names into a single feature
     funders = build_entity_query(
         "funders",
-        "funder_ids",
-        "funder_names",
+        "funders.ror",
+        "funders.name",
         dmp.funding,
-        lambda fund: fund.funder.id,
+        lambda fund: fund.funder.ror,
         lambda fund: fund.funder.name,
     )
     if funders is not None:
         should.append(funders)
+
+    # Awards
+    awards = build_awards_query(
+        "awards",
+        dmp.external_data.awards,
+    )
+    if awards is not None:
+        should.append(awards)
 
     # Title and abstract
     has_text = dmp.title is not None or dmp.abstract is not None
@@ -290,7 +309,7 @@ def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) 
         should.append(
             {
                 "more_like_this": {
-                    "_name": name_group("content"),
+                    "_name": "content",
                     "fields": ["title", "abstract"],
                     "like": content,
                     "min_term_freq": 1,
@@ -311,7 +330,7 @@ def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) 
             },
         }
     ]
-    if dmp.project_start is not None:
+    if dmp.project_start is not None and dmp.project_start >= pendulum.date(1990, 1, 1):
         gte = dmp.project_start.format("YYYY-MM-DD")
         lte = dmp.project_end.add(years=project_end_buffer_years).format("YYYY-MM-DD")
         filters.append(
@@ -346,6 +365,7 @@ def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) 
                 "title": {
                     "type": "fvh",
                     "number_of_fragments": 0,
+                    "fragment_size": 0,
                 },
                 "abstract": {
                     "type": "fvh",
@@ -367,7 +387,7 @@ def build_query(dmp: DMPModel, max_results: int, project_end_buffer_years: int) 
 
 
 def build_entity_query(
-    group_name: str,
+    path: str,
     id_field: str,
     name_field: str,
     items: list,
@@ -385,7 +405,7 @@ def build_entity_query(
             entity_queries.append(
                 {
                     "constant_score": {
-                        "_name": name_value(group_name, idx, "id", entity_id),
+                        "_name": id_field,
                         "filter": {"term": {id_field: entity_id}},
                         "boost": 2,
                     }
@@ -396,31 +416,83 @@ def build_entity_query(
             entity_queries.append(
                 {
                     "constant_score": {
-                        "_name": name_value(group_name, idx, "name", entity_name),
+                        "_name": name_field,
                         "filter": {"match_phrase": {name_field: {"query": entity_name, "slop": 3}}},
                         "boost": 1,
                     }
                 }
             )
 
-        if entity_queries:
+        if len(entity_queries) > 1:
             should_queries.append(
                 {
                     "dis_max": {
-                        "_name": name_entity(group_name, idx),
                         "tie_breaker": 0,
                         "queries": entity_queries,
                     },
                 }
             )
+        elif len(entity_queries) == 1:
+            should_queries.append(entity_queries[0])
 
     if should_queries:
         return {
-            "bool": {
-                "_name": name_group(group_name),
-                "minimum_should_match": 1,
-                "should": should_queries,
+            "nested": {
+                "path": path,
+                "query": {
+                    "bool": {
+                        "minimum_should_match": 1,
+                        "should": should_queries,
+                    }
+                },
+                "inner_hits": {"name": path},
+            },
+        }
+
+    return None
+
+
+def build_awards_query(
+    path: str,
+    awards: list[Award],
+) -> Optional[dict]:
+    """The dis_max ensures that all the variants of a single award only contribute
+    a maximum score of 10."""
+
+    award_queries = []
+    for award in awards:
+        queries = []
+        for award_id in award.award_id.all_variants:
+            queries.append(
+                {
+                    "constant_score": {
+                        "_name": "awards.award_id",
+                        "filter": {"term": {"awards.award_id": award_id}},
+                        "boost": 10,
+                    }
+                }
+            )
+        award_queries.append(
+            {
+                "dis_max": {
+                    "tie_breaker": 0,
+                    "queries": queries,
+                }
             }
+        )
+
+    if len(award_queries):
+        return {
+            "nested": {
+                "path": path,
+                "query": {
+                    "bool": {
+                        "minimum_should_match": 1,
+                        "should": award_queries,
+                    }
+                },
+                "inner_hits": {"name": path},
+            },
         }
 
     return None