+++ queries/dmpworks/python/dmpworks/opensearch/enrich_dmps.py
@@ -71,14 +71,14 @@ def enrich_dmps(
                 for fund in dmp.funding:
                     # Parse Award IDs, which can be found in both funding_opportunity_id
                     # and award_id
-                    award_ids = parse_award_text(fund.funder.id, fund.funding_opportunity_id)
-                    award_ids.extend(parse_award_text(fund.funder.id, fund.award_id))
+                    award_ids = parse_award_text(fund.funder.ror, fund.funding_opportunity_id)
+                    award_ids.extend(parse_award_text(fund.funder.ror, fund.award_id))
                     award_ids = set(award_ids)
 
                     # Fetch additional data for each award ID
                     for award_id in award_ids:
                         dois = fetch_funded_dois(award_id, email=email)
-                        awards.append(Award(funder=fund.funder, award_id=award_id, funded_dois=dois, award_url=""))
+                        awards.append(Award(funder=fund.funder, award_id=award_id, funded_dois=dois))
 
                 log.debug(f"Save additional metadata for DMP: {dmp.doi}")
                 external_data = ExternalData(updated=pendulum.now(tz="UTC"), awards=awards).model_dump()